Renamed some test files
