+++ shared/src/iosTest/kotlin/co/touchlab/kmp/PlatformTest.kt
@@ -1,0 +1,11 @@
+package co.touchlab.kampstarter
+
+import kotlinx.coroutines.runBlocking
+
+actual fun <T> runTest(block: suspend () -> T) { runBlocking { block() } }
+
+class SqlDelightTestJvm : SqlDelightTest()
+
+class KtorTestJvm: KtorTest()
+
+class BreedModelTestJvm: BreedModelTest()
